SEF/src/sef/ATestTask/SecondActivity/SecondActivityTest.java
package sef.ATestTask.SecondActivity;
// Complete Code
import junit.framework.TestCase;
import org.junit.Test;

public class SecondActivityTest extends TestCase {


	//TODO try to achieve 100% of test coverage

	private SecondActivity theCalculator;

	protected void setUp() throws Exception {
		super.setUp();
		//	Initialize variables to be used here
		theCalculator = new SecondActivity();
	}

	protected void tearDown() throws Exception {
		super.tearDown();
	}

	//Example ->
	@Test
	public void testGetSumPositiveValue() {
		assertEquals(15, (theCalculator.getSum(5, 10)));
		assertEquals(15, (theCalculator.getSum(10, 5)));
	}

	//Example ->
	@Test
	public void testGetSumNegativeValue() {
		assertEquals(5, (theCalculator.getSum(-5, 10)));
		assertEquals(-5, (theCalculator.getSum(5, -10)));
	}
	@Test
public void  sum() {
		assertEquals(15.2, (theCalculator.sum(10.2,5)));
		assertEquals(15.2, (theCalculator.sum(5,10.2)));
}
}
@Test
public void divide() {
	assertEquals(3.03, theCalculator.divide(100,33));
	assertEquals(0.303, (theCalculator.divide(33,100)));
}


